# serve/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HypothesisRequest": {
        "additionalProperties": false,
        "properties": {
          "head_entity_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Head Entity Id",
            "type": "string"
          },
          "tail_entity_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Tail Entity Id",
            "type": "string"
          }
        },
        "required": [
          "head_entity_id",
          "tail_entity_id"
        ],
        "title": "HypothesisRequest",
        "type": "object"
      },
      "HypothesisResponse": {
        "additionalProperties": false,
        "properties": {
          "head_entity_id": {
            "title": "Head Entity Id",
            "type": "string"
          },
          "hypotheses": {
            "items": {
              "$ref": "#/components/schemas/HypothesisItemResponse"
            },
            "title": "Hypotheses",
            "type": "array"
          },
          "library_id": {
            "title": "Library Id",
            "type": "string"
          },
          "tail_entity_id": {
            "title": "Tail Entity Id",
            "type": "string"
          }
        },
        "required": [
          "library_id",
          "head_entity_id",
          "tail_entity_id",
          "hypotheses"
        ],
        "title": "HypothesisResponse",
        "type": "object"
      },
...
      "LibraryCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "library_id": {
            "maxLength": 31,
            "minLength": 3,
            "title": "Library Id",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "library_id",
          "name"
        ],
        "title": "LibraryCreateRequest",
        "type": "object"
      },
      "LibraryResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "library_id": {
            "title": "Library Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "library_id",
          "name",
          "description",
          "created_at"
        ],
        "title": "LibraryResponse",
        "type": "object"
      },
      "LibrarySchemaResponse": {
        "additionalProperties": false,
        "properties": {
          "entity_types": {
            "items": {
              "$ref": "#/components/schemas/TypeColorResponse"
            },
            "title": "Entity Types",
            "type": "array"
          },
          "library_id": {
            "title": "Library Id",
            "type": "string"
          },
          "relation_types": {
            "items": {
              "type": "string"
            },
            "title": "Relation Types",
            "type": "array"
          }
        },
        "required": [
          "library_id",
          "entity_types",
          "relation_types"
        ],
        "title": "LibrarySchemaResponse",
        "type": "object"
      },
...
      "NeighborhoodResponse": {
        "additionalProperties": false,
        "properties": {
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "library_id": {
            "title": "Library Id",
            "type": "string"
          },
          "triples": {
            "items": {
              "$ref": "#/components/schemas/TripleResponse"
            },
            "title": "Triples",
            "type": "array"
          }
        },
        "required": [
          "library_id",
          "entity_id",
          "depth",
          "triples"
        ],
        "title": "NeighborhoodResponse",
        "type": "object"
      },
      "QARequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "QARequest",
        "type": "object"
      },
      "QAResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Citations",
            "type": "array"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "library_id": {
            "title": "Library Id",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "library_id",
          "question",
          "answer",
          "citations",
          "model",
          "input_tokens",
          "output_tokens",
          "duration_ms"
        ],
        "title": "QAResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "RAG-KG Copilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/libraries": {
      "get": {
        "operationId": "list_libraries_v1_libraries_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LibraryResponse"
                  },
                  "title": "Response List Libraries V1 Libraries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Libraries",
        "tags": [
          "libraries"
        ]
      },
      "post": {
        "operationId": "create_library_v1_libraries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Library",
        "tags": [
          "libraries"
        ]
      }
    },
    "/v1/libraries/{library_id}": {
      "delete": {
        "operationId": "delete_library_v1_libraries__library_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Library",
        "tags": [
          "libraries"
        ]
      },
      "get": {
        "operationId": "get_library_v1_libraries__library_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Library",
        "tags": [
          "libraries"
        ]
      }
    },
    "/v1/libraries/{library_id}/entities/{entity_id}/neighborhood": {
      "get": {
        "operationId": "entity_neighborhood_v1_libraries__library_id__entities__entity_id__neighborhood_get",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Depth",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborhoodResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Entity Neighborhood",
        "tags": [
          "libraries"
        ]
      }
    },
    "/v1/libraries/{library_id}/hypothesis": {
      "post": {
        "operationId": "generate_hypothesis_v1_libraries__library_id__hypothesis_post",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HypothesisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HypothesisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Hypothesis",
        "tags": [
          "libraries"
        ]
      }
    },
...
    "/v1/libraries/{library_id}/qa": {
      "post": {
        "operationId": "qa_v1_libraries__library_id__qa_post",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QAResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Qa",
        "tags": [
          "libraries"
        ]
      }
    },
    "/v1/libraries/{library_id}/qa/stream": {
      "get": {
        "description": "SSE stream: meta → tokens → citations → done. (Token chunking is\nclient-side fan-out for now; will become real LLM streaming when the\nLLM adapter exposes streaming completions.)",
        "operationId": "qa_stream_v1_libraries__library_id__qa_stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "question",
            "required": true,
            "schema": {
              "title": "Question",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Qa Stream",
        "tags": [
          "libraries"
        ]
      }
    },
...
    "/v1/libraries/{library_id}/schema": {
      "get": {
        "operationId": "library_schema_v1_libraries__library_id__schema_get",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibrarySchemaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Library Schema",
        "tags": [
          "libraries"
        ]
      }
    },
    "/v1/libraries/{library_id}/stats": {
      "get": {
        "operationId": "library_stats_v1_libraries__library_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Library Stats",
        "tags": [
          "libraries"
        ]
      }
    }
  }
}